+++ _test_annotations/lib/test_annotations.dart
@@ -1,3 +1,8 @@
 class TestAnnotation {
   const TestAnnotation();
 }
+
+class TestAnnotationWithNew {
+  final String message;
+  const new({this.message = 'default'});
+}